[Refactor] 코드 스타일 통일
