+++ src/components/SmallButton/SmallButton.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import CreateIcon from "@/assets/create.svg?react";
+import CreateIcon from "@/assets/icons/create.svg?react";
 import { SmallButtonProps } from "./SmallButton.types";
 import { base, variants, disabledStyle } from "./SmallButton.styled";
 import { cn } from "@/utils/cn";
